rename some file and added normal delete button components
